# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Arodona Jewelry Backend API",
    "version": "1.0.0",
    "description": "E-commerce API for Arodona Jewelry store"
  },
  "servers": [
    {
      "url": "http://localhost:5858",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/products": {
      "post": {
        "summary": "Create a new product",
        "tags": ["Products"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCreateResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all products",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/favorites/toggle": {
      "post": {
        "summary": "Toggle favorite for a product",
        "tags": ["Favorites"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Favorite toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteToggleResponse"
                }
              }
            }
          },
          "201": {
            "description": "Favorite added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteToggleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/favorites": {
      "get": {
        "summary": "List user favorites",
        "tags": ["Favorites"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorites retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/favorites/check/{product_id_or_slug}": {
      "get": {
        "summary": "Check if product is favorited",
        "tags": ["Favorites"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "product_id_or_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite status checked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteCheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/favorites/{favorite_id}": {
      "delete": {
        "summary": "Remove a favorite",
        "tags": ["Favorites"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "favorite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteRemoveResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["name", "description", "price"]
      },
...
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryInfo"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rating_avg": {
            "type": "number"
          },
          "rating_count": {
            "type": "integer"
          },
          "ratings_breakdown": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "name", "description", "price", "created_at", "updated_at"]
      },
      "CategoryInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": ["id", "name", "slug"]
      },
      "ProductCreateResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/ProductResponse"
          }
        },
        "required": ["message", "product"]
      },
      "ProductListResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "skip": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            }
          }
        },
        "required": ["total", "skip", "limit", "products"]
      },
...
      "FavoriteToggleRequest": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string"
          }
        },
        "required": ["product"]
      },
      "FavoriteToggleResponse": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": ["added", "removed", "exists"]
          },
          "product_id": {
            "type": "string"
          },
          "favorite_id": {
            "type": "string"
          }
        },
        "required": ["action", "product_id"]
      },
      "FavoriteCheckResponse": {
        "type": "object",
        "properties": {
          "favorited": {
            "type": "boolean"
          }
        },
        "required": ["favorited"]
      },
      "FavoriteListResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "favorite_id": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "product": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "required": ["favorite_id", "created_at", "product"]
            }
          }
        },
        "required": ["count", "items"]
      },
      "FavoriteRemoveResponse": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": ["removed"]
          },
          "favorite_id": {
            "type": "string"
          }
        },
        "required": ["action", "favorite_id"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}